Refactor Learnly workspace into modular presenter
