+++ src/ui/views/browser/components/learnly/constants.js
@@ -1,0 +1,26 @@
+export const VIEW_CATALOG = 'catalog';
+export const VIEW_DETAIL = 'detail';
+export const VIEW_MY_COURSES = 'myCourses';
+export const VIEW_PRICING = 'pricing';
+export const VIEW_FREE = 'freeCourses';
+
+export const CATEGORY_DEFINITIONS = [
+  { id: 'writing', label: 'Writing & Storycraft', skills: ['writing'] },
+  { id: 'audience', label: 'Audience Engagement', skills: ['audience'] },
+  { id: 'promotion', label: 'Promotion & Funnels', skills: ['promotion'] },
+  { id: 'research', label: 'Market Research', skills: ['research'] },
+  { id: 'visualEditing', label: 'Visual & Editing Skills', skills: ['visual', 'editing'] },
+  { id: 'commerce', label: 'Commerce & Fulfillment', skills: ['commerce'] },
+  { id: 'technical', label: 'Technical Skills (Software, Infra)', skills: ['software', 'infrastructure'] },
+  { id: 'audio', label: 'Audio / Media Skills', skills: ['audio'] },
+  { id: 'general', label: 'Multidisciplinary', skills: [] }
+];
+
+export default {
+  VIEW_CATALOG,
+  VIEW_DETAIL,
+  VIEW_MY_COURSES,
+  VIEW_PRICING,
+  VIEW_FREE,
+  CATEGORY_DEFINITIONS
+};
